+++ builtin/logical/mysql/path_role_create.go
@@ -50,18 +50,31 @@ func (b *backend) pathRoleCreateRead(
 		lease = &configLease{}
 	}
 
-	// Generate our username and password. MySQL limits user to 16 characters
+	// Generate our username and password. The username will be a
+	// concatenation of:
+	//
+	// - the role name, truncated to role.rolenameLength (default 4)
+	// - the token display name, truncated to role.displaynameLength (default 4)
+	// - a UUID
+	//
+	// the entire contactenated string is then truncated to role.usernameLength,
+	// which by default is 16 due to limitations in older but still-prevalant
+	// versions of MySQL.
+	roleName := name
+	if len(roleName) > role.RolenameLength {
+		roleName = roleName[:role.RolenameLength]
+	}
 	displayName := req.DisplayName
-	if len(displayName) > 10 {
-		displayName = displayName[:10]
+	if len(displayName) > role.DisplaynameLength {
+		displayName = displayName[:role.DisplaynameLength]
 	}
 	userUUID, err := uuid.GenerateUUID()
 	if err != nil {
 		return nil, err
 	}
-	username := fmt.Sprintf("%s-%s", displayName, userUUID)
-	if len(username) > 16 {
-		username = username[:16]
+	username := fmt.Sprintf("%s-%s-%s", roleName, displayName, userUUID)
+	if len(username) > role.UsernameLength {
+		username = username[:role.UsernameLength]
 	}
 	password, err := uuid.GenerateUUID()
 	if err != nil {